Arborescence modifiée
